Rename LieAlgebraOptimizer to RiemannianGradientOptimizer (#4153)
* Create RiemannianGradientOptimizer

* Update docstring

* Add tests

* update optimization and numpy docs

* update versions and deprecations

* lint new tests

* better generator syntax

* Replace deprecated by renamed in changelogs


* Remove old name from interfaces docs

---------
