+++ doc/development/deprecations.rst
@@ -6,6 +6,11 @@ Deprecations
 Pending deprecations
 --------------------
 
+* ``LieAlgebraOptimizer`` is renamed. Please use ``RiemannianGradientOptimizer`` instead.
+  
+  - Deprecated in v0.31
+  - Will be removed in v0.32
+
 * The ``Operation.base_name`` property is deprecated. Please use ``Operator.name`` or ``type(obj).__name__`` instead.
 
   - Still accessible in v0.31
