+++ doc/releases/changelog-dev.md
@@ -176,10 +176,13 @@
 
 <h3>Deprecations 👋</h3>
 
+* `LieAlgebraOptimizer` is renamed. Please use `RiemannianGradientOptimizer` instead.
+  [(#4153)(https://github.com/PennyLaneAI/pennylane/pull/4153)]
+
 * `Operation.base_name` is deprecated. Please use `Operation.name` or `type(op).__name__` instead.
 
-* ``QuantumScript``'s ``name`` keyword argument and property are deprecated.
-  This also affects ``QuantumTape`` and ``OperationRecorder``.
+* `QuantumScript`'s `name` keyword argument and property are deprecated.
+  This also affects `QuantumTape` and `OperationRecorder`.
   [(#4141)](https://github.com/PennyLaneAI/pennylane/pull/4141)
 
 * `qml.grouping` module is removed. The functionality has been reorganized in the `qml.pauli` module.
