+++ pennylane/optimize/__init__.py
@@ -24,6 +24,7 @@ from .gradient_descent import GradientDescentOptimizer
 from .lie_algebra import LieAlgebraOptimizer
 from .momentum import MomentumOptimizer
 from .nesterov_momentum import NesterovMomentumOptimizer
+from .riemannian_gradient import RiemannianGradientOptimizer
 from .rms_prop import RMSPropOptimizer
 from .qng import QNGOptimizer
 from .rotosolve import RotosolveOptimizer
@@ -43,6 +44,7 @@ __all__ = [
     "NesterovMomentumOptimizer",
     "RMSPropOptimizer",
     "QNGOptimizer",
+    "RiemannianGradientOptimizer",
     "RotosolveOptimizer",
     "RotoselectOptimizer",
     "ShotAdaptiveOptimizer",
