+++ pennylane/optimize/lie_algebra.py
@@ -14,130 +14,11 @@
 """Lie algebra optimizer"""
 import warnings
 
-import numpy as np
-from scipy.sparse.linalg import expm
-import pennylane as qml
+from .riemannian_gradient import RiemannianGradientOptimizer
 
 
-@qml.qfunc_transform
-def append_time_evolution(tape, riemannian_gradient, t, n, exact=False):
-    r"""Append an approximate time evolution, corresponding to a Riemannian
-    gradient on the Lie group, to an existing circuit.
-
-    We want to implement the time evolution generated by an operator of the form
-
-    .. math::
-
-        \text{grad} f(U) = sum_i c_i O_i,
-
-    where :math:`O_i` are Pauli words and :math:`c_t \in \mathbb{R}`.
-    If ``exact`` is ``False``, we Trotterize this operator and apply the unitary
-
-    .. math::
-
-        U' = \prod_{n=1}^{N_{Trot.}} \left(\prod_i \exp{-it / N_{Trot.} O_i}\right),
-
-    which is then appended to the current circuit.
-
-    If ``exact`` is ``True``, we calculate the exact time evolution for the Riemannian gradient
-     by way of the matrix exponential.
-
-    .. math:
-
-        U' = \exp{-it \text{grad} f(U)}
-
-    and append this unitary.
-
-    Args:
-        tape (QuantumTape or .QNode): circuit to transform.
-        riemannian_gradient (.Hamiltonian): Hamiltonian object representing the Riemannian gradient.
-        t (float): time evolution parameter.
-        n (int): number of Trotter steps.
-
-    """
-    for obj in tape.operations:
-        qml.apply(obj)
-    if exact:
-        qml.QubitUnitary(
-            expm(-1j * t * riemannian_gradient.sparse_matrix().toarray()),
-            wires=range(len(riemannian_gradient.wires)),
-        )
-    else:
-        qml.templates.ApproxTimeEvolution(riemannian_gradient, t, n)
-
-    for obj in tape.measurements:
-        qml.apply(obj)
-
-
-@qml.batch_transform
-def algebra_commutator(tape, observables, lie_algebra_basis_names, nqubits):
-    """Calculate the Riemannian gradient in the Lie algebra with the parameter shift rule
-    (see :meth:`LieAlgebraOptimizer.get_omegas`).
-
-    Args:
-        tape (.QuantumTape or .QNode): input circuit
-        observables (list[.Observable]): list of observables to be measured. Can be grouped.
-        lie_algebra_basis_names (list[str]): List of strings corresponding to valid Pauli words.
-        nqubits (int): the number of qubits.
-
-    Returns:
-        function or tuple[list[QuantumTape], function]:
-
-        - If the input is a QNode, an object representing the Riemannian gradient function
-          of the QNode that can be executed with the same arguments as the QNode to obtain
-          the Lie algebra commutator.
-
-        - If the input is a tape, a tuple containing a
-          list of generated tapes, together with a post-processing
-          function to be applied to the results of the evaluated tapes
-          in order to obtain the Lie algebra commutator.
-    """
-    tapes_plus_total = []
-    tapes_min_total = []
-    for obs in observables:
-        for o in obs:
-            # create a list of tapes for the plus and minus shifted circuits
-            queues_plus = [qml.queuing.AnnotatedQueue() for _ in lie_algebra_basis_names]
-            queues_min = [qml.queuing.AnnotatedQueue() for _ in lie_algebra_basis_names]
-
-            # loop through all operations on the input tape
-            for op in tape.operations:
-                for t in queues_plus + queues_min:
-                    with t:
-                        qml.apply(op)
-            for i, t in enumerate(queues_plus):
-                with t:
-                    qml.PauliRot(
-                        np.pi / 2,
-                        lie_algebra_basis_names[i],
-                        wires=list(range(nqubits)),
-                    )
-                    qml.expval(o)
-            for i, t in enumerate(queues_min):
-                with t:
-                    qml.PauliRot(
-                        -np.pi / 2,
-                        lie_algebra_basis_names[i],
-                        wires=list(range(nqubits)),
-                    )
-                    qml.expval(o)
-            tapes_plus_total.extend(
-                [
-                    qml.tape.QuantumScript(*qml.queuing.process_queue(q))
-                    for q, p in zip(queues_plus, lie_algebra_basis_names)
-                ]
-            )
-            tapes_min_total.extend(
-                [
-                    qml.tape.QuantumScript(*qml.queuing.process_queue(q))
-                    for q, p in zip(queues_min, lie_algebra_basis_names)
-                ]
-            )
-    return tapes_plus_total + tapes_min_total, None
-
-
-class LieAlgebraOptimizer:
-    r"""Lie algebra optimizer.
+class LieAlgebraOptimizer(RiemannianGradientOptimizer):
+    r"""Lie algebra optimizer. Deprecated in favour of :class:`~.RiemannianGradientOptimizer`.
 
     Riemannian gradient descent algorithms can be used to optimize a function directly on a Lie group
     as opposed to on a Euclidean parameter space. Consider the function
@@ -241,141 +122,15 @@ class LieAlgebraOptimizer:
     # pylint: disable=too-many-arguments
     # pylint: disable=too-many-instance-attributes
     def __init__(self, circuit, stepsize=0.01, restriction=None, exact=False, trottersteps=1):
-        if not isinstance(circuit, qml.QNode):
-            raise TypeError(f"circuit must be a QNode, received {type(circuit)}")
-
-        self.circuit = circuit
-        self.circuit.construct([], {})
-        self.hamiltonian = circuit.func().obs
-        if not isinstance(self.hamiltonian, qml.Hamiltonian):
-            raise TypeError(
-                f"circuit must return the expectation value of a Hamiltonian,"
-                f"received {type(circuit.func().obs)}"
-            )
-        self.nqubits = len(circuit.device.wires)
-
-        if self.nqubits > 4:
-            warnings.warn(
-                "The exact Riemannian gradient is exponentially expensive in the number of qubits, "
-                f"optimizing a {self.nqubits} qubit circuit may be slow.",
-                UserWarning,
-            )
-        if restriction is not None and not isinstance(restriction, qml.Hamiltonian):
-            raise TypeError(f"restriction must be a Hamiltonian, received {type(restriction)}")
-        (
-            self.lie_algebra_basis_ops,
-            self.lie_algebra_basis_names,
-        ) = self.get_su_n_operators(restriction)
-        self.exact = exact
-        self.trottersteps = trottersteps
-        self.coeffs, self.observables = self.hamiltonian.terms()
-        self.stepsize = stepsize
-
-    def step(self):
-        r"""Update the circuit with one step of the optimizer.
-
-        Returns:
-           float: the optimized circuit and the objective function output prior
-           to the step.
-        """
-        return self.step_and_cost()[0]
-
-    def step_and_cost(self):
-        r"""Update the circuit with one step of the optimizer and return the corresponding
-        objective function value prior to the step.
-
-        Returns:
-            tuple[.QNode, float]: the optimized circuit and the objective function output prior
-            to the step.
-        """
-        # pylint: disable=not-callable
-
-        cost = self.circuit()
-        omegas = self.get_omegas()
-        non_zero_omegas = -omegas[omegas != 0]
-
-        nonzero_idx = np.nonzero(omegas)[0]
-        non_zero_lie_algebra_elements = [self.lie_algebra_basis_names[i] for i in nonzero_idx]
-
-        lie_gradient = qml.Hamiltonian(
-            non_zero_omegas,
-            [qml.pauli.string_to_pauli_word(ps) for ps in non_zero_lie_algebra_elements],
+        warnings.warn(
+            "``LieAlgebraOptimizer`` is deprecated. "
+            "Please, use ``RiemannianGradientOptimizer`` instead.",
+            UserWarning,
         )
-        new_circuit = append_time_evolution(
-            lie_gradient, self.stepsize, self.trottersteps, self.exact
-        )(self.circuit.func)
-
-        # we can set diff_method=None because the gradient of the QNode is computed
-        # directly in this optimizer
-        self.circuit = qml.QNode(new_circuit, self.circuit.device, diff_method=None)
-        return self.circuit, cost
-
-    def get_su_n_operators(self, restriction):
-        r"""Get the SU(N) operators. The dimension of the group is :math:`N^2-1`.
-
-        Args:
-            restriction (.Hamiltonian): Restrict the Riemannian gradient to a subspace.
-
-        Returns:
-            tuple[list[array[complex]], list[str]]: list of :math:`N^2 \times N^2` NumPy complex arrays and corresponding Pauli words.
-        """
-
-        operators = []
-        names = []
-        # construct the corresponding pennylane observables
-        wire_map = dict(zip(range(self.nqubits), range(self.nqubits)))
-        if restriction is None:
-            for ps in qml.pauli.pauli_group(self.nqubits):
-                operators.append(ps)
-                names.append(qml.pauli.pauli_word_to_string(ps, wire_map=wire_map))
-        else:
-            for ps in set(restriction.ops):
-                operators.append(ps)
-                names.append(qml.pauli.pauli_word_to_string(ps, wire_map=wire_map))
-
-        return operators, names
-
-    def get_omegas(self):
-        r"""Measure the coefficients of the Riemannian gradient with respect to a Pauli word basis.
-        We want to calculate the components of the Riemannian gradient in the Lie algebra
-        with respect to a Pauli word basis. For a Hamiltonian of the form :math:`H = \sum_i c_i O_i`,
-        where :math:`c_i\in\mathbb{R}`, this can be achieved by calculating
-
-        .. math::
-
-            \omega_{i,j} = \text{Tr}(c_i[\rho, O_i] P_j)
-
-        where :math:`P_j` is a Pauli word in the set of Pauli monomials on :math:`N` qubits.
-
-        Via the parameter shift rule, the commutator can be calculated as
-
-        .. math::
-
-            [\rho, O_i] = \frac{1}{2}(V(\pi/2) \rho V^\dagger(\pi/2) - V(-\pi/2) \rho V^\dagger(-\pi/2))
-
-        where :math:`V` is the unitary generated by the Pauli word :math:`V(\theta) = \exp\{-i\theta P_j\}`.
-
-        Returns:
-            array: array of omegas for each direction in the Lie algebra.
-        """
-
-        obs_groupings, _ = qml.pauli.group_observables(self.observables, self.coeffs)
-        # get all circuits we need to calculate the coefficients
-        circuits = algebra_commutator(
-            self.circuit.qtape,
-            obs_groupings,
-            self.lie_algebra_basis_names,
-            self.nqubits,
-        )[0]
-        circuits = qml.execute(circuits, self.circuit.device, gradient_fn=None)
-        circuits_plus = np.array(circuits[: len(circuits) // 2]).reshape(
-            len(self.coeffs), len(self.lie_algebra_basis_names)
+        super().__init__(
+            circuit,
+            stepsize=stepsize,
+            restriction=restriction,
+            exact=exact,
+            trottersteps=trottersteps,
         )
-        circuits_min = np.array(circuits[len(circuits) // 2 :]).reshape(
-            len(self.coeffs), len(self.lie_algebra_basis_names)
-        )
-
-        # For each observable O_i in the Hamiltonian, we have to calculate all Lie coefficients
-        omegas = circuits_plus - circuits_min
-
-        return np.dot(self.coeffs, omegas)
