+++ tests/optimize/test_riemannian_gradient_optimizer.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """
-Unit tests for the ``LieAlgebraOptimizer``.
+Unit tests for the ``RiemannianGradientOptimizer``.
 """
 import pytest
 
@@ -20,7 +20,7 @@ from scipy.sparse.linalg import expm
 import numpy as np
 
 import pennylane as qml
-from pennylane.optimize import LieAlgebraOptimizer
+from pennylane.optimize import RiemannianGradientOptimizer
 
 
 def circuit_1():
@@ -67,24 +67,6 @@ hamiltonian_3 = qml.Hamiltonian(
     coeffs=[-2.0], observables=[qml.PauliY(0) @ qml.PauliY(1) @ qml.PauliY(2)]
 )
 
-circuits = [circuit_1, circuit_2, circuit_3]
-hamiltonians = [hamiltonian_1, hamiltonian_2, hamiltonian_3]
-
-
-@pytest.mark.parametrize("circuit, hamiltonian", zip(circuits, hamiltonians))
-def test_lie_algebra_deprecation_warning(circuit, hamiltonian):
-    """Test that creating a ``LieAlgebraOptimizer`` outputs the correct deprecation warning."""
-    nqubits = max([max(ps.wires) for ps in hamiltonian.ops]) + 1
-    dev = qml.device("default.qubit", wires=nqubits)
-
-    @qml.qnode(dev)
-    def lie_circuit():
-        circuit()
-        return qml.expval(hamiltonian)
-
-    with pytest.warns(UserWarning, match="``LieAlgebraOptimizer`` is deprecated."):
-        _ = LieAlgebraOptimizer(circuit=lie_circuit)
-
 
 @pytest.mark.parametrize(
     "circuit,hamiltonian",
@@ -96,11 +78,11 @@ def test_lie_algebra_deprecation_warning(circuit, hamiltonian):
         (circuit_3, hamiltonian_3),
     ],
 )
-def test_lie_algebra_omegas(circuit, hamiltonian):
+def test_riemannian_gradient_omegas(circuit, hamiltonian):
     """Test that we calculate the Riemannian gradient coefficients Tr{[rho, H] P_j} correctly."""
     # pylint: disable=no-member
 
-    nqubits = max([max(ps.wires) for ps in hamiltonian.ops]) + 1
+    nqubits = max(max(ps.wires) for ps in hamiltonian.ops) + 1
     wires = range(nqubits)
     dev = qml.device("default.qubit", wires=nqubits)
 
@@ -110,20 +92,20 @@ def test_lie_algebra_omegas(circuit, hamiltonian):
         return qml.state()
 
     @qml.qnode(dev)
-    def lie_circuit():
+    def test_circuit():
         circuit()
         return qml.expval(hamiltonian)
 
     phi = get_state()
     rho = np.outer(phi, phi.conj())
     hamiltonian_np = hamiltonian.sparse_matrix(wire_order=wires).toarray()
-    lie_algebra_np = hamiltonian_np @ rho - rho @ hamiltonian_np
-    opt = LieAlgebraOptimizer(circuit=lie_circuit)
+    riemannian_gradient_np = hamiltonian_np @ rho - rho @ hamiltonian_np
+    opt = RiemannianGradientOptimizer(circuit=test_circuit)
     ops = opt.get_su_n_operators(None)[0]
     omegas_np = []
     for op in ops:
         op = qml.math.expand_matrix(op.matrix(), op.wires, wires)
-        omegas_np.append(1j * np.trace(lie_algebra_np @ op))
+        omegas_np.append(1j * np.trace(riemannian_gradient_np @ op))
     omegas = opt.get_omegas()
     assert np.allclose(omegas, omegas_np)
 
@@ -138,10 +120,10 @@ def test_lie_algebra_omegas(circuit, hamiltonian):
         (circuit_3, hamiltonian_3),
     ],
 )
-def test_lie_algebra_omegas_restricted(circuit, hamiltonian):
+def test_riemannian_gradient_omegas_restricted(circuit, hamiltonian):
     """Test that we calculate the (restricted) Riemannian gradient coefficients correctly."""
     # pylint: disable=no-member
-    nqubits = max([max(ps.wires) for ps in hamiltonian.ops]) + 1
+    nqubits = max(max(ps.wires) for ps in hamiltonian.ops) + 1
     wires = range(nqubits)
     dev = qml.device("default.qubit", wires=nqubits)
 
@@ -151,26 +133,26 @@ def test_lie_algebra_omegas_restricted(circuit, hamiltonian):
         return qml.state()
 
     @qml.qnode(dev)
-    def lie_circuit():
+    def test_circuit():
         circuit()
         return qml.expval(hamiltonian)
 
     phi = get_state()
     rho = np.outer(phi, phi.conj())
     hamiltonian_np = hamiltonian.sparse_matrix(wire_order=wires).toarray()
-    lie_algebra_np = hamiltonian_np @ rho - rho @ hamiltonian_np
+    riemannian_gradient_np = hamiltonian_np @ rho - rho @ hamiltonian_np
 
     restriction = qml.Hamiltonian(
         coeffs=[1.0] * 3,
         observables=[qml.PauliX(0), qml.PauliY(1), qml.PauliY(0) @ qml.PauliY(1)],
     )
 
-    opt = LieAlgebraOptimizer(circuit=lie_circuit, restriction=restriction)
+    opt = RiemannianGradientOptimizer(circuit=test_circuit, restriction=restriction)
     ops = opt.get_su_n_operators(restriction)[0]
     omegas_np = []
     for op in ops:
         op = qml.math.expand_matrix(op.matrix(), op.wires, wires)
-        omegas_np.append(1j * np.trace(lie_algebra_np @ op))
+        omegas_np.append(1j * np.trace(riemannian_gradient_np @ op))
     omegas = opt.get_omegas()
 
     assert np.allclose(omegas, omegas_np)
@@ -185,10 +167,10 @@ def test_lie_algebra_omegas_restricted(circuit, hamiltonian):
         (circuit_3, hamiltonian_3),
     ],
 )
-def test_lie_algebra_evolution(circuit, hamiltonian):
+def test_riemannian_gradient_evolution(circuit, hamiltonian):
     """Test that the optimizer produces the correct unitary to append."""
     # pylint: disable=no-member
-    nqubits = max([max(ps.wires) for ps in hamiltonian.ops]) + 1
+    nqubits = max(max(ps.wires) for ps in hamiltonian.ops) + 1
     wires = range(nqubits)
     dev = qml.device("default.qubit", wires=nqubits)
 
@@ -198,18 +180,18 @@ def test_lie_algebra_evolution(circuit, hamiltonian):
         return qml.state()
 
     @qml.qnode(dev)
-    def lie_circuit():
+    def test_circuit():
         circuit()
         return qml.expval(hamiltonian)
 
     phi = get_state()
     rho = np.outer(phi, phi.conj())
     hamiltonian_np = hamiltonian.sparse_matrix(wire_order=wires).toarray()
-    lie_algebra_np = hamiltonian_np @ rho - rho @ hamiltonian_np
+    riemannian_gradient_np = hamiltonian_np @ rho - rho @ hamiltonian_np
 
-    phi_exact = expm(-0.1 * lie_algebra_np * 2**nqubits) @ phi
+    phi_exact = expm(-0.1 * riemannian_gradient_np * 2**nqubits) @ phi
     rho_exact = np.outer(phi_exact, phi_exact.conj())
-    opt = LieAlgebraOptimizer(circuit=lie_circuit, stepsize=0.1, exact=True)
+    opt = RiemannianGradientOptimizer(circuit=test_circuit, stepsize=0.1, exact=True)
     opt.step_and_cost()
     cost_pl = opt.circuit()
     cost_exact = np.trace(rho_exact @ hamiltonian_np)
@@ -226,18 +208,18 @@ def test_lie_algebra_evolution(circuit, hamiltonian):
         (circuit_3, hamiltonian_3),
     ],
 )
-def test_lie_algebra_step(circuit, hamiltonian):
+def test_riemannian_gradient_step(circuit, hamiltonian):
     """Test that we can take subsequent steps with the optimizer."""
-    nqubits = max([max(ps.wires) for ps in hamiltonian.ops]) + 1
+    nqubits = max(max(ps.wires) for ps in hamiltonian.ops) + 1
 
     dev = qml.device("default.qubit", wires=nqubits)
 
     @qml.qnode(dev)
-    def lie_circuit():
+    def test_circuit():
         circuit()
         return qml.expval(hamiltonian)
 
-    opt = LieAlgebraOptimizer(circuit=lie_circuit)
+    opt = RiemannianGradientOptimizer(circuit=test_circuit)
     opt.step()
     opt.step()
 
@@ -252,33 +234,33 @@ def test_lie_algebra_step(circuit, hamiltonian):
         (circuit_3, hamiltonian_3),
     ],
 )
-def test_lie_algebra_step_trotterstep(circuit, hamiltonian):
+def test_riemannian_gradient_step_trotterstep(circuit, hamiltonian):
     """Test that we can take subsequent steps with the optimizer."""
-    nqubits = max([max(ps.wires) for ps in hamiltonian.ops]) + 1
+    nqubits = max(max(ps.wires) for ps in hamiltonian.ops) + 1
 
     dev = qml.device("default.qubit", wires=nqubits)
 
     @qml.qnode(dev)
-    def lie_circuit():
+    def test_circuit():
         circuit()
         return qml.expval(hamiltonian)
 
-    opt = LieAlgebraOptimizer(circuit=lie_circuit, trottersteps=3)
+    opt = RiemannianGradientOptimizer(circuit=test_circuit, trottersteps=3)
     opt.step()
     opt.step()
 
 
-def test_lie_algebra_circuit_input_1_check():
+def test_riemannian_gradient_circuit_input_1_check():
     """Test that a type error is raise for non-QNode circuits."""
 
     def circuit():
         qml.RY(0.5, wires=0)
 
     with pytest.raises(TypeError, match="circuit must be a QNode"):
-        LieAlgebraOptimizer(circuit=circuit, stepsize=0.001)
+        RiemannianGradientOptimizer(circuit=circuit, stepsize=0.001)
 
 
-def test_lie_algebra_hamiltonian_input_1_check():
+def test_riemannian_gradient_hamiltonian_input_1_check():
     """Test that a type error is raise for non-QNode circuits."""
 
     @qml.qnode(qml.device("default.qubit", wires=3))
@@ -290,10 +272,10 @@ def test_lie_algebra_hamiltonian_input_1_check():
         TypeError,
         match="circuit must return the expectation value of a Hamiltonian",
     ):
-        LieAlgebraOptimizer(circuit=circuit, stepsize=0.001)
+        RiemannianGradientOptimizer(circuit=circuit, stepsize=0.001)
 
 
-def test_lie_algebra_nqubits_check():
+def test_riemannian_gradient_nqubits_check():
     """Test that we warn if the system is too big."""
 
     @qml.qnode(qml.device("default.qubit", wires=5))
@@ -302,10 +284,10 @@ def test_lie_algebra_nqubits_check():
         return qml.expval(qml.Hamiltonian(coeffs=[-1.0], observables=[qml.PauliX(0)]))
 
     with pytest.warns(UserWarning, match="The exact Riemannian gradient is exponentially"):
-        LieAlgebraOptimizer(circuit=circuit, stepsize=0.001)
+        RiemannianGradientOptimizer(circuit=circuit, stepsize=0.001)
 
 
-def test_lie_algebra_restriction_check():
+def test_riemannian_gradient_restriction_check():
     """Test that a type error is raise for non-QNode circuits."""
 
     @qml.qnode(qml.device("default.qubit", wires=3))
@@ -318,7 +300,7 @@ def test_lie_algebra_restriction_check():
         TypeError,
         match="restriction must be a Hamiltonian",
     ):
-        LieAlgebraOptimizer(circuit=circuit, restriction=restriction, stepsize=0.001)
+        RiemannianGradientOptimizer(circuit=circuit, restriction=restriction, stepsize=0.001)
 
 
 def test_docstring_example():
@@ -336,7 +318,7 @@ def test_docstring_example():
         qml.RY(0.6, wires=[0])
         return qml.expval(hamiltonian)
 
-    opt = LieAlgebraOptimizer(circuit=quant_fun, stepsize=0.1)
+    opt = RiemannianGradientOptimizer(circuit=quant_fun, stepsize=0.1)
 
     for _ in range(12):
         circuit, cost = opt.step_and_cost()
@@ -360,7 +342,7 @@ def test_docstring_example_exact():
         qml.RY(0.6, wires=[0])
         return qml.expval(hamiltonian)
 
-    opt = LieAlgebraOptimizer(circuit=quant_fun, stepsize=0.1, exact=True)
+    opt = RiemannianGradientOptimizer(circuit=quant_fun, stepsize=0.1, exact=True)
 
     for _ in range(12):
         circuit, cost = opt.step_and_cost()
@@ -383,7 +365,7 @@ def test_example_shots():
         qml.RY(0.6, wires=[0])
         return qml.expval(hamiltonian)
 
-    opt = LieAlgebraOptimizer(circuit=quant_fun, stepsize=0.1, exact=False)
+    opt = RiemannianGradientOptimizer(circuit=quant_fun, stepsize=0.1, exact=False)
 
     for _ in range(3):
         opt.step_and_cost()
